+++ marketplace/services/vtex/private/products/service.py
@@ -31,12 +31,13 @@ Example:
         # Use products data as needed
 """
 
-from typing import List
+from typing import List, Optional
 
 from marketplace.services.vtex.exceptions import CredentialsValidationError
 from marketplace.services.vtex.utils.data_processor import DataProcessor
 from marketplace.services.vtex.business.rules.rule_mappings import RULE_MAPPINGS
 from marketplace.services.vtex.utils.data_processor import FacebookProductDTO
+from marketplace.wpp_products.models import Catalog
 
 
 class PrivateProductsService:
@@ -60,14 +61,38 @@ class PrivateProductsService:
         self.check_is_valid_domain(domain)
         return self.client.is_valid_credentials(domain)
 
-    def list_all_products(self, domain, catalog) -> List[FacebookProductDTO]:
-        config = catalog.vtex_app.config
-        active_sellers = self.client.list_active_sellers(domain)
-        skus_ids = self.client.list_all_products_sku_ids(domain)
+    def list_all_products(
+        self,
+        domain: str,
+        config: dict,
+        catalog: Catalog,
+        sellers: Optional[List[str]] = None,
+        update_product=False,
+    ) -> List[FacebookProductDTO]:
+        active_sellers = set(self.client.list_active_sellers(domain))
+        if sellers is not None:
+            valid_sellers = [seller for seller in sellers if seller in active_sellers]
+            invalid_sellers = set(sellers) - active_sellers
+            if invalid_sellers:
+                print(
+                    f"Warning: Sellers IDs {invalid_sellers} are not active and will be ignored."
+                )
+            sellers_ids = valid_sellers
+        else:
+            sellers_ids = list(active_sellers)
+
+        skus_ids = self.client.list_all_active_products(domain)
         rules = self._load_rules(config.get("rules", []))
         store_domain = config.get("store_domain")
         products_dto = self.data_processor.process_product_data(
-            skus_ids, active_sellers, self, domain, store_domain, rules, catalog
+            skus_ids,
+            sellers_ids,
+            self,
+            domain,
+            store_domain,
+            rules,
+            update_product,
+            catalog,
         )
         return products_dto
 
@@ -80,20 +105,19 @@ class PrivateProductsService:
         )  # TODO: Change to pvt_simulate_cart_for_seller
 
     def update_webhook_product_info(
-        self, domain: str, skus_ids: list, catalog
+        self, domain: str, skus_ids: list, seller_ids: list, catalog: Catalog
     ) -> List[FacebookProductDTO]:
         config = catalog.vtex_app.config
-        seller_ids = self.client.list_active_sellers(domain)
         rules = self._load_rules(config.get("rules", []))
         store_domain = config.get("store_domain")
         updated_products_dto = self.data_processor.process_product_data(
-            skus_ids,
-            seller_ids,
-            self,
-            domain,
-            store_domain,
-            rules,
-            catalog,
+            skus_ids=skus_ids,
+            active_sellers=seller_ids,
+            service=self,
+            domain=domain,
+            store_domain=store_domain,
+            rules=rules,
+            catalog=catalog,
             update_product=True,
         )