+++ marketplace/wpp_products/tasks.py
@@ -7,21 +7,28 @@ from celery import shared_task
 from django.db import reset_queries, close_old_connections
 
 from marketplace.clients.facebook.client import FacebookClient
-from marketplace.services.webhook.vtex.webhook_manager import WebhookQueueManager
-from marketplace.wpp_products.models import Catalog
+
+from marketplace.wpp_products.models import (
+    Catalog,
+    ProductUploadLog,
+    UploadProduct,
+    WebhookLog,
+)
 from marketplace.clients.flows.client import FlowsClient
 from marketplace.celery import app as celery_app
 from marketplace.services.vtex.generic_service import (
     ProductUpdateService,
     ProductInsertionService,
     VtexServiceBase,
+    ProductInsertionBySellerService,
 )
 from marketplace.services.vtex.generic_service import APICredentials
 from marketplace.core.types import APPTYPES
 from marketplace.applications.models import App
 
 from django_redis import get_redis_connection
-from django.core.cache import cache
+
+from marketplace.wpp_products.utils import ProductUploader
 
 
 logger = logging.getLogger(__name__)
@@ -120,10 +127,10 @@ class FacebookCatalogSyncService:
 def task_insert_vtex_products(**kwargs):
     print("Starting task: 'task_insert_vtex_products'")
     vtex_service = ProductInsertionService()
-    # flows_service = FlowsService(FlowsClient())
 
     credentials = kwargs.get("credentials")
     catalog_uuid = kwargs.get("catalog_uuid")
+    sellers = kwargs.get("sellers")
 
     if not all([credentials, catalog_uuid]):
         logger.error(
@@ -145,7 +152,7 @@ def task_insert_vtex_products(**kwargs):
             domain=credentials["domain"],
         )
         print(f"Starting first product insert for catalog: {str(catalog.name)}")
-        products = vtex_service.first_product_insert(api_credentials, catalog)
+        products = vtex_service.first_product_insert(api_credentials, catalog, sellers)
         if products is None:
             print("There are no products to be shipped after processing the rules")
             return
@@ -158,21 +165,6 @@ def task_insert_vtex_products(**kwargs):
     finally:
         close_old_connections()
 
-    # Temporarily removes the sending of products to flows [04-02-2024]
-    # try:
-    #     dict_catalog = {
-    #         "name": catalog.name,
-    #         "facebook_catalog_id": catalog.facebook_catalog_id,
-    #     }
-    #     flows_service.update_vtex_products(
-    #         products, str(catalog.app.flow_object_uuid), dict_catalog
-    #     )
-    #     print("Products successfully sent to flows")
-    # except Exception as e:
-    #     logger.error(
-    #         f"Error on send vtex products to flows for catalog {catalog_uuid}, {e}"
-    #     )
-
     print(
         f"finishing creation products, task: 'task_insert_vtex_products' catalog {catalog.name}"
     )
@@ -181,91 +173,239 @@ def task_insert_vtex_products(**kwargs):
 
 @celery_app.task(name="task_update_vtex_products")
 def task_update_vtex_products(**kwargs):
-    print("Starting task: 'task_update_vtex_products'")
+    start_time = datetime.now()
     vtex_base_service = VtexServiceBase()
-    # flows_service = FlowsService(FlowsClient())
 
     app_uuid = kwargs.get("app_uuid")
+    webhook = kwargs.get("webhook")
 
-    queue_manager = WebhookQueueManager(app_uuid)
-    redis = get_redis_connection()
-    lock_key = queue_manager.get_lock_key()
+    sku_id = webhook.get("IdSku")
+    seller_an = webhook.get("An")
+    seller_chain = webhook.get("SellerChain")
+    try:
+        current_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f")
+        logger.info(
+            f"Processing product update for App UUID: {app_uuid}, "
+            f"SKU_ID: {sku_id} at {current_time}. "
+            f"'An':{seller_an}, 'SellerChain': {seller_chain}."
+        )
+        vtex_app = App.objects.get(uuid=app_uuid, configured=True, code="vtex")
+        (
+            domain,
+            app_key,
+            app_token,
+        ) = vtex_base_service.get_vtex_credentials_or_raise(vtex_app)
+        api_credentials = APICredentials(
+            app_key=app_key, app_token=app_token, domain=domain
+        )
 
-    lock = redis.lock(lock_key, timeout=7200)
-    if lock.acquire(blocking=False):
-        processing_key = queue_manager.get_sku_list_key()
-        try:
-            while True:
-                webhooks_in_processing = cache.get(processing_key, [])
-                if not webhooks_in_processing:
-                    print("There are no products to deal with in cache.")
-                    break
-
-                skus_ids = queue_manager.dequeue_webhook_data()
-                if not skus_ids:
-                    print("There are no products in dequeue process.")
-                    break
-
-                current_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f")
-                print(
-                    f"Processing product update for App UUID: {app_uuid}, SKU IDs: {skus_ids} at {current_time}"
-                )
-                vtex_app = App.objects.get(uuid=app_uuid, configured=True, code="vtex")
-                (
-                    domain,
-                    app_key,
-                    app_token,
-                ) = vtex_base_service.get_vtex_credentials_or_raise(vtex_app)
-                api_credentials = APICredentials(
-                    app_key=app_key, app_token=app_token, domain=domain
-                )
-
-                catalog = vtex_app.vtex_catalogs.first()
-                if not catalog or not catalog.feeds.first():
-                    logger.error(
-                        f"No data feed found in the database. Vtex app: {vtex_app.uuid}"
-                    )
-                    continue
-
-                product_feed = catalog.feeds.first()
-                print(f"Starting product update for app: {str(vtex_app.uuid)}")
-
-                vtex_update_service = ProductUpdateService(
-                    api_credentials, catalog, skus_ids, product_feed
-                )
-                products = vtex_update_service.webhook_product_insert()
-                if products is None:
-                    print(
-                        f"No products to process after treatment for VTEX app {app_uuid}. Task ending."
-                    )
-                    continue
-
-                # Temporarily removes the sending of products to flows [03-30-2024]
-
-                # dict_catalog = {
-                #     "name": catalog.name,
-                #     "facebook_catalog_id": catalog.facebook_catalog_id,
-                # }
-
-                # try:
-                #     flows_service.update_vtex_products(
-                #         products, str(catalog.app.flow_object_uuid), dict_catalog
-                #     )
-                #     print("Products successfully sent to flows")
-                # except Exception as e:
-                #     logger.error(f"Failed to send products to flows: {str(e)}")
-
-                close_old_connections()
-
-        except Exception as e:
-            logger.error(
-                f"An error occurred during the updating Webhook vtex products for app {app_uuid}, {str(e)}"
+        catalog = vtex_app.vtex_catalogs.first()
+        if not catalog or not catalog.feeds.first():
+            logger.info(
+                f"No data feed found in the database. Vtex app: {vtex_app.uuid}"
             )
+            return
+
+        product_feed = catalog.feeds.first()
+
+        vtex_update_service = ProductUpdateService(
+            api_credentials, catalog, [sku_id], product_feed, webhook
+        )
+        products = vtex_update_service.webhook_product_insert()
+        if products is None:
+            logger.info(
+                f"No products to process after treatment for VTEX app {app_uuid}. Task ending."
+            )
+            return
+
+        close_old_connections()
+        # Webhook Log
+        WebhookLog.objects.create(sku_id=sku_id, data=webhook, vtex_app=vtex_app)
+
+    except Exception as e:
+        logger.error(
+            f"An error occurred during the updating Webhook vtex products for app {app_uuid}, {str(e)}"
+        )
+
+    end_time = datetime.now()
+    duration = (end_time - start_time).total_seconds()
+    minutes, seconds = divmod(duration, 60)
+
+    logger.info(
+        f"Finishing process update vtex product to SKU:{sku_id} App: {app_uuid}"
+    )
+    logger.info(f"Task completed in {int(minutes)} minutes and {int(seconds)} seconds.")
+
+    redis_client = get_redis_connection()
+    lock_key = f"upload_lock:{app_uuid}"
+    # Check for existing upload lock for the vtex app
+    if not redis_client.exists(lock_key):
+        print(f"No active upload task for App: {app_uuid}, starting upload.")
+        celery_app.send_task(
+            "task_upload_vtex_products",
+            kwargs={"app_vtex_uuid": app_uuid},
+            queue="vtex-product-upload",
+        )
+    else:
+        print(f"An upload task is already in progress for App: {app_uuid}.")
+
+    print("=" * 40)
+
+
+@celery_app.task(name="task_forward_vtex_webhook")
+def task_forward_vtex_webhook(**kwargs):
+    app_uuid = kwargs.get("app_uuid")
+    webhook = kwargs.get("webhook")
+
+    try:
+        app = App.objects.get(uuid=app_uuid, configured=True, code="vtex")
+    except App.DoesNotExist:
+        logger.info(f"No VTEX App configured with the provided UUID: {app_uuid}")
+        return
+
+    can_synchronize = app.config.get("initial_sync_completed", False)
+
+    celery_queue = app.config.get("celery_queue_name", "product_synchronization")
+
+    if not can_synchronize:
+        print(f"Initial sync not completed. App:{str(app.uuid)}")
+        return
+
+    sku_id = webhook.get("IdSku")
+
+    if not sku_id:
+        raise ValueError(f"SKU ID not provided in the request. App:{str(app.uuid)}")
+
+    celery_app.send_task(
+        "task_update_vtex_products",
+        kwargs={"app_uuid": str(app_uuid), "webhook": webhook},
+        queue=celery_queue,
+        ignore_result=True,
+    )
+
+
+@celery_app.task(name="task_upload_vtex_products")
+def task_upload_vtex_products(**kwargs):
+    app_vtex_uuid = kwargs.get("app_vtex_uuid")
+    app_vtex = App.objects.get(uuid=app_vtex_uuid)
+    redis_client = get_redis_connection()
+    lock_key = f"upload_lock:{app_vtex_uuid}"
+    lock_expiration_time = 15 * 60  # 15 minutes
+
+    # Attempt to acquire the lock
+    if redis_client.set(lock_key, "locked", nx=True, ex=lock_expiration_time):
+        try:
+            catalogs = app_vtex.vtex_catalogs.all()
+            if not catalogs.exists():
+                print("No catalogs found.")
+                return
+
+            for catalog in catalogs:
+                if catalog.feeds.first():
+                    print(f"Processing upload for catalog: {catalog.name}")
+                    uploader = ProductUploader(catalog=catalog)
+                    uploader.process_and_upload(
+                        redis_client, lock_key, lock_expiration_time
+                    )
 
         finally:
-            lock.release()
+            # Release the lock
+            redis_client.delete(lock_key)
     else:
-        print("Unable to acquire lock, another process is running.")
+        print(f"Upload task for App: {app_vtex_uuid} is already in progress.")
 
-    print(f"Finishing update vtex product to App: {app_uuid}")
+    print(f"Processing upload for App: {app_vtex_uuid}")
+
+
+@celery_app.task(name="task_cleanup_vtex_logs_and_uploads")
+def task_cleanup_vtex_logs_and_uploads():
+    # Delete all records from the ProductUploadLog and WebhookLog tables
+    ProductUploadLog.objects.all().delete()
+    WebhookLog.objects.all().delete()
+
+    # Delete all UploadProduct records with "success" status
+    UploadProduct.objects.filter(status="success").delete()
+
+    print("Logs and successful uploads have been cleaned up.")
+
+
+def send_sync(app_uuid: str, webhook: dict):
+    try:
+        app = App.objects.get(uuid=app_uuid, configured=True, code="vtex")
+    except App.DoesNotExist:
+        logger.info(f"No VTEX App configured with the provided UUID: {app_uuid}")
+        return
+
+    can_synchronize = app.config.get("initial_sync_completed", False)
+
+    if not can_synchronize:
+        print(f"Initial sync not completed. App:{str(app.uuid)}")
+        return
+
+    celery_queue = app.config.get("celery_queue_name", "product_synchronization")
+    sku_id = webhook.get("IdSku")
+
+    if not sku_id:
+        raise ValueError(f"SKU ID not provided in the request. App:{str(app.uuid)}")
+
+    celery_app.send_task(
+        "task_update_vtex_products",
+        kwargs={"app_uuid": str(app_uuid), "webhook": webhook},
+        queue=celery_queue,
+        ignore_result=True,
+    )
+
+
+@celery_app.task(name="task_insert_vtex_products_by_sellers")
+def task_insert_vtex_products_by_sellers(**kwargs):
+    print("Starting insertion products by seller")
+    vtex_service = ProductInsertionBySellerService()
+
+    credentials = kwargs.get("credentials")
+    catalog_uuid = kwargs.get("catalog_uuid")
+    sellers = kwargs.get("sellers")
+
+    if not sellers:
+        logger.error(
+            "Missing required parameters [seller] for task_insert_vtex_products_by_sellers"
+        )
+        return
+
+    if not all([credentials, catalog_uuid]):
+        logger.error(
+            "Missing required parameters [credentials, catalog_uuid] for task_insert_vtex_products"
+        )
+        return
+
+    try:
+        # Reset queries and close old connections for a clean state and performance.
+        # Prevents memory leak from stored queries and unstable database connections
+        reset_queries()
+        close_old_connections()
+
+        catalog = Catalog.objects.get(uuid=catalog_uuid)
+        api_credentials = APICredentials(
+            app_key=credentials["app_key"],
+            app_token=credentials["app_token"],
+            domain=credentials["domain"],
+        )
+        print(
+            f"Starting 'insertion_products_by_seller' for catalog: {str(catalog.name)}"
+        )
+        products = vtex_service.insertion_products_by_seller(
+            api_credentials, catalog, sellers
+        )
+        if products is None:
+            print("There are no products to be shipped after processing the rules.")
+            return
+
+    except Exception as e:
+        logger.exception(
+            f"An error occurred during the 'insertion_products_by_seller' for catalog {catalog.name}, {e}"
+        )
+        return
+    finally:
+        close_old_connections()
+
+    print(f"finishing 'insertion_products_by_seller'catalog {catalog.name}")
     print("=" * 40)